DERBY-6884: SYSCS_IMPORT_TABLE_LOBS_FROM_EXTFILE can't import lob data
This change modifies the ImportLobFile.getString() and
ImportReadData.initExternalLobFile() methods so that they use a Java "long"
variable for the offset into the external lob file; prior to this change
they were using a Java "int" variable and hence would malfunction when the
lob offsets exceeded Integer.MAX_VALUE ( 2,147,483,647 ).

The regression test which demonstrates these problems is a bit slow to run;
on my system, it takes approximately 15 minutes to execute, and requires
about 10 GB of available disk space during the test run. Therefore, the
test cases are placed in a new test program (Derby6884Test), which is not
listed in the "standard" system test suites, but rather is only added to the
"largedata" suite. The new test can also be run by itself, e.g.:

ant -Dderby.junit.testclass=org.apache.derbyTesting.functionTests.tests.largedata.Derby6884Test junit-clean junit-single


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1742057 13f79535-47bb-0310-9956-ffa450edef68

+++ java/engine/org/apache/derby/impl/load/ImportLobFile.java
@@ -128,10 +128,10 @@ class ImportLobFile
      * @param length  length of the the data.
      * @exception  IOException  on any I/O error.     
      */
-    public String getString(int offset, int length) throws IOException {
+    public String getString(long offset, int length) throws IOException {
         lobInputStream.seek(offset);
         lobLimitIn.clearLimit();
-        lobLimitIn.setLimit((int) length);
+        lobLimitIn.setLimit(length);
         
         // wrap a reader on top of the stream, so that calls 
         // to read the clob data from the file can read the
